Moved PyBot related functions out -- will be going to the DolaBot repo at https://github.com/kjhf/DolaBot
Moved battlefy-toolkit functionality out -- will be going to the IPL repo at https://github.com/IPLSplatoon/Battlefy-Toolkit
