--- PyBot/constants/bot_constants.py
@@ -1,1 +1,0 @@
-COMMAND_PREFIX = '~'
