--- PyBot/constants/emojis.py
@@ -1,35 +1,0 @@
-# To find these, use \:emojiName: in the Discord chat.
-BATTLEFY = '<:battlefy:810346162161844274>'
-DISCORD = '<:discord:810348495079866370>'
-TWITCH = '<:twitch:810346162023432213>'
-TWITTER = '<:twitter:810346162418614312>'
-TROPHY = '🏆'
-CROWN = '👑'
-
-# Abilities
-ABILITY_DOUBLER = "<:AbilityDoubler:841052609648525382>"
-BOMB_DEFENSE_UP_DX = "<:BombDefenseUpDX:841052609333821471>"
-COMEBACK = "<:Comeback:841052609300922419>"
-DROP_ROLLER = "<:DropRoller:841052609859289118>"
-HAUNT = "<:Haunt:841052609850900530>"
-INK_RECOVERY_UP = "<:InkRecoveryUp:841052609866629120>"
-INK_RESISTANCE_UP = "<:InkResistanceUp:841052609627684875>"
-INK_SAVER_MAIN = "<:InkSaverMain:841052609875148911>"
-INK_SAVER_SUB = "<:InkSaverSub:841052609992327168>"
-LAST_DITCH_EFFORT = "<:LastDitchEffort:841052609870692362>"
-MAIN_POWER_UP = "<:MainPowerUp:841052609972011008>"
-NINJA_SQUID = "<:NinjaSquid:841052610122219574>"
-OBJECT_SHREDDER = "<:ObjectShredder:841052609858895914>"
-OPENING_GAMBIT = "<:OpeningGambit:841052610080669697>"
-QUICK_RESPAWN = "<:QuickRespawn:841052610256437289>"
-QUICK_SUPER_JUMP = "<:QuickSuperJump:841052609723891733>"
-RESPAWN_PUNISHER = "<:RespawnPunisher:841052610050916383>"
-RUN_SPEED_UP = "<:RunSpeedUp:841052610320400475>"
-SPECIAL_CHARGE_UP = "<:SpecialChargeUp:841052610193784862>"
-SPECIAL_POWER_UP = "<:SpecialPowerUp:841052610199027764>"
-SPECIAL_SAVER = "<:SpecialSaver:841052610038726667>"
-STEALTH_JUMP = "<:StealthJump:841052610131656724>"
-SUB_POWER_UP = "<:SubPowerUp:841052609912897567>"
-SWIM_SPEED_UP = "<:SwimSpeedUp:841052610106097674>"
-TENACITY = "<:Tenacity:841052610173206548>"
-THERMAL_INK = "<:ThermalInk:841052609803583489>"
