--- PyBot/entry/main.py
@@ -1,6 +1,0 @@
-from PyBot.entry.DolaBot import DolaBot
-
-if __name__ == '__main__':
-    dola = DolaBot()
-    dola.do_the_thing()
-    print("Main exited!")
