--- vlee/battlefyToolkit/__init__.py
@@ -1,1 +1,0 @@
-from .org import get_org, tournament_ids
