--- vlee/requirements.txt
@@ -1,3 +1,0 @@
-requests
-pymongo
-python-dateutil
